🚚 Move TabIcons into Asset struct
