--- HRHN/Sources/Utils/Design/TabIcons/TabIcons.swift
@@ -1,17 +1,0 @@
-//
-//  TabIcons.swift
-//  HRHN
-//
-//  Created by 민채호 on 2023/02/10.
-//
-
-import Foundation
-
-enum TabbarIcons: String {
-    case todaySelected = "Today_selected"
-    case todayUnselected = "Today_unselected"
-    case recordSelected = "Record_selected"
-    case recordUnselected = "Record_unselected"
-    case storageSelected = "Storage_selected"
-    case storageUnselected = "Storage_unselected"
-}
